# openapi.yaml
openapi: 3.0.3
info:
  title: Raptor API
  description: |
    Self-hosted webhook/email/DNS capture & inspection. This document is the
    spec-first source of truth for the versioned management API under `/api/v1`.
    The SPA is a pure client of it.

    **Authentication.** When `--require-auth` is enabled, send `Api-Key: <uuid>`.
    With no key configured the API is open (first-run bootstrap mode).
  version: "1.0"
servers:
  - url: /api/v1
paths:
...
  /tokens/{tokenID}/requests/{requestID}/response:
    parameters:
      - $ref: "#/components/parameters/TokenID"
      - $ref: "#/components/parameters/RequestID"
    post:
      summary: Supply the response for a pending request (CLI listen flow)
      operationId: setResponse
      description: >
        When the token has `listen > 0`, a capture is held open until this
        endpoint supplies a response. Body may be base64 (`content`) or raw
        (`content_text`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: integer }
                content: { type: string, description: "base64-encoded body" }
                content_text: { type: string, description: "raw body (alternative to content)" }
                headers: { type: object, additionalProperties: { type: string } }
      responses:
        "204": { description: Response delivered }
        "404": { description: No request is waiting (not listening / already responded) }
...
components:
  parameters:
    TokenID:
      name: tokenID
      in: path
      required: true
      description: Token UUID or alias
      schema: { type: string }
    RequestID:
      name: requestID
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: Api-Key